feat: update landing page assets and restructure components for improved clarity
- Increased logo size in README for better visibility.
- Removed unused favicon.ico and several landing components to streamline the project.
- Refactored ButtonClickSound import path for consistency.
- Cleaned up code by deleting obsolete animation and layout components.
