+++ trx-web/src/components/hero/dino-path.ts
