--- trx-web/src/components/landing/animations.ts
@@ -1,11 +1,0 @@
-export const ease = [0.25, 0.1, 0.25, 1] as const;
-
-export const itemVariants = {
-  hidden:  { opacity: 0, y: 16 },
-  visible: { opacity: 1, y: 0, transition: { duration: 0.4, ease } },
-};
-
-export const staggerVariants = {
-  hidden:  {},
-  visible: { transition: { staggerChildren: 0.08, delayChildren: 0.05 } },
-};
