--- trx-web/src/components/landing/icons.tsx
@@ -1,13 +1,0 @@
-// UI / feature icons: Lucide set via react-icons
-export { LuSearch as IconSearch } from "react-icons/lu";
-export { LuBox as IconBox } from "react-icons/lu";
-export { LuLayers as IconLayers } from "react-icons/lu";
-export { LuZap as IconZap } from "react-icons/lu";
-export { LuClipboard as IconClipboard } from "react-icons/lu";
-export { LuCheck as IconCheck } from "react-icons/lu";
-
-// Brand / social icons: Simple Icons set via react-icons
-export { SiGithub as IconGithub } from "react-icons/si";
-export { SiApple as IconApple } from "react-icons/si";
-export { SiArchlinux as IconArch } from "react-icons/si";
-export { SiDebian as IconLinux } from "react-icons/si";
